minor improvements

+++ lib/runtime/zycap_linux.c
@@ -151,7 +151,7 @@ int Zycap_Prefetch_Bitstream    ( char * bs_name, t_bitstream * bitstream)
     }
     else
     {
-        printf("[Zycap] bitstream prefetch %s: has already an address: %x \n", bs_name, (uint32_t)bs_name);
+        printf("[Zycap] bitstream prefetch %s: has already an address: %x \n", bs_name, bs_name);
     }
     int bytes_read = fread(bitstream->data,1,size,fp);
     bitstream->size = size;
